pyproject.toml
[project]
name = "tmaze-undistort"
version = "0.1.0"
description = "T-Maze video undistortion pipeline for camera calibration and rectification"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "matplotlib>=3.5.0",
    "numpy>=1.20.0",
    "opencv-python>=4.5.0",
    "pyyaml>=5.4.0",
    "scikit-image>=0.18.0",
    "shapely>=2.0.0",
    "sleap-io>=0.5.0",
    "tqdm>=4.60.0",
    "imageio>=2.9.0",
]
authors = [
    {name = "Leo", email = "leo@example.com"}
]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
]

[project.optional-dependencies]
dev = [
    "jupyterlab>=4.0.0",
    "pandas>=2.0.0",
    "pytest>=7.0.0",
]

[project.scripts]
tmaze-undistort = "tmaze_undistort.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/tmaze_undistort"]
